refactor: remove classic backoffice organizations view
Delete the legacy /organizations-classic backoffice module now that
organizations_v2 is the primary interface. Relocate the still-used shared
helpers (analytics, orders_import, and 8 forms) into organizations_v2,
drop the classic-only endpoints, account_review components, schemas, and
forms, remove the "Switch to Classic View" links, and repoint the
users-page delete action to the v2 delete_dialog route.
