+++ server/polar/backoffice/organizations_v2/endpoints.py
@@ -28,20 +28,6 @@ from polar.account_credit.repository import AccountCreditRepository
 from polar.account_credit.service import account_credit_service
 from polar.auth.scope import READ_ONLY_SCOPES
 from polar.auth.service import auth as auth_service
-from polar.backoffice.organizations.analytics import (
-    OrganizationSetupAnalyticsService,
-    PaymentAnalyticsService,
-)
-from polar.backoffice.organizations.forms import (
-    AddPaymentMethodDomainForm,
-    OrganizationOrdersImportForm,
-    UpdateOrganizationBasicForm,
-    UpdateOrganizationDetailsForm,
-    UpdateOrganizationInternalNotesForm,
-    UpdateOrganizationSocialsForm,
-    UpdateRateLimitGroupForm,
-)
-from polar.backoffice.organizations.orders_import import orders_import_sse
 from polar.config import settings
 from polar.enums import PayoutAccountType
 from polar.file.repository import FileRepository
@@ -123,7 +109,21 @@ from ..layout import layout
 from ..responses import HXRedirectResponse
 from ..toast import add_toast
 from . import priority as review_priority
-from .forms import UpdateAccountSettingsForm
+from .analytics import (
+    OrganizationSetupAnalyticsService,
+    PaymentAnalyticsService,
+)
+from .forms import (
+    AddPaymentMethodDomainForm,
+    OrganizationOrdersImportForm,
+    UpdateAccountSettingsForm,
+    UpdateOrganizationBasicForm,
+    UpdateOrganizationDetailsForm,
+    UpdateOrganizationInternalNotesForm,
+    UpdateOrganizationSocialsForm,
+    UpdateRateLimitGroupForm,
+)
+from .orders_import import orders_import_sse
 from .priority import Signals
 from .views.detail_view import OrganizationDetailView
 from .views.list_view import (
